+++ infrastructure-android/src/main/kotlin/AndroidApplicationPlugin.kt
@@ -3,7 +3,6 @@ package com.redmadrobot.build
 import com.android.build.gradle.AppExtension
 import com.redmadrobot.build.extension.RedmadrobotExtension
 import com.redmadrobot.build.internal.android
-import com.redmadrobot.build.internal.configureKotlinDependencies
 import org.gradle.api.Project
 
 public class AndroidApplicationPlugin : BaseAndroidPlugin() {
@@ -11,7 +10,6 @@ public class AndroidApplicationPlugin : BaseAndroidPlugin() {
     override fun Project.configure() {
         applyBaseAndroidPlugin("com.android.application")
 
-        configureKotlinDependencies(redmadrobotExtension.kotlinVersion, "implementation")
         configureApp(redmadrobotExtension)
     }
 }