+++ infrastructure/src/main/kotlin/internal/Kotlin.kt
@@ -4,8 +4,6 @@ import com.redmadrobot.build.InfrastructurePlugin
 import com.redmadrobot.build.extension.isRunningOnCi
 import com.redmadrobot.build.kotlinCompile
 import org.gradle.api.Project
-import org.gradle.kotlin.dsl.dependencies
-import org.gradle.kotlin.dsl.kotlin
 
 public fun InfrastructurePlugin.configureKotlin() {
     val warningsAsErrors = project.getWarningsAsErrorsProperty()
@@ -21,10 +19,4 @@ public fun InfrastructurePlugin.configureKotlin() {
 /** Use property value if it exists or fallback to true if running on CI. */
 private fun Project.getWarningsAsErrorsProperty(): Boolean {
     return findBooleanProperty("warningsAsErrors") ?: isRunningOnCi
-}
-
-public fun InfrastructurePlugin.configureKotlinDependencies(kotlinVersion: String, configuration: String = api) {
-    project.dependencies {
-        configuration(kotlin("stdlib-jdk8", version = kotlinVersion))
-    }
 }