+++ infrastructure/src/main/kotlin/internal/Test.kt
@@ -1,23 +1,7 @@
 package com.redmadrobot.build.internal
 
-import com.redmadrobot.build.InfrastructurePlugin
 import com.redmadrobot.build.extension.TestOptions
 import org.gradle.api.tasks.testing.Test
-import org.gradle.kotlin.dsl.dependencies
-import org.gradle.kotlin.dsl.kotlin
-
-public fun InfrastructurePlugin.configureKotlinTestDependencies(kotlinVersion: String, testOptions: TestOptions) {
-    project.dependencies {
-        val kotlinJunitModule = if (testOptions.useJunitPlatform) {
-            "test-junit5"
-        } else {
-            "test-junit"
-        }
-
-        testImplementation(kotlin("test", version = kotlinVersion))
-        testImplementation(kotlin(kotlinJunitModule, version = kotlinVersion))
-    }
-}
 
 public fun Test.setTestOptions(testOptions: TestOptions) {
     if (testOptions.useJunitPlatform) {